+++ angular-workspace/projects/ni/nimble-angular/src/public-api.ts
@@ -10,6 +10,8 @@ export * from './directives/breadcrumb-item/nimble-breadcrumb-item.directive';
 export * from './directives/breadcrumb-item/nimble-breadcrumb-item.module';
 export * from './directives/button/nimble-button.directive';
 export * from './directives/button/nimble-button.module';
+export * from './directives/card-button/nimble-card-button.directive';
+export * from './directives/card-button/nimble-card-button.module';
 export * from './directives/checkbox/nimble-checkbox-control-value-accessor.directive';
 export * from './directives/checkbox/nimble-checkbox.directive';
 export * from './directives/checkbox/nimble-checkbox.module';